pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "viral-snake"
version = "0.1.0"
description = "Snakemake pipeline for viral metagenomics"
authors = [{name = "Dmitry Fedorov", email = "fedorov.de@gmail.com"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "snakemake>=7.0.0",
    "pyyaml",
    "pandas",
]

[tool.setuptools]
packages = ["viral_snake"]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
]